Reformat parser grammar
In an effort to see better what the JSON parser grammar is doing,
we reduced the length of the grammar lines, making it easier to
read it in a window (of 132 chars wide) along side other windows
(where say, debug output found or a debugger is running).

In an effort to understand which lines of C code are doing
what in a `{ ... }` C-code block, the C-code block is now divided
via comments, into 3 sections:

```c
/* pre action debugging */
...
/* action */
...
/* post-action debugging */
```

In an effort to understand what the JSON parser is doing while
it is executing, `json_dbg()` calls within `{ ... }` C-code blocks
and been make more consistent in terms of what they print,
and more compact in terms of line length.

In an effort to understand what the JSON parser is doing while
it is executing, the debug level for `json_dbg()` calls has
been divided into:

DBG_LOW - starting, action and, ending debug statements
DBG_LED - struct json * types,
          and where applicable ugly_text and ugly_length

The actual grammar was NOT changed.  Only some of the `json_dbg()`
calls were changed.

TODO: Try out our idea about on how to address the 4
      `XXX - need more code here - XXX` sections.