Use ugly_leng instead of ugly_length in jparse.y
It appears that ugly_leng (technically yyleng) is a global variable
after all so use it instead of the ugly_length we had. This will
probably be dealt with better when we look into making the parser and
scanner re-entrant but for now this works and it removes another
variable to keep track of.