Check if JSON input is all whitespace
Whatever I was thinking when changing this some days back is no longer
valid and this way we don't even call the parser on only whitespace
(which is invalid json). The idea I had I think has to do with NUL bytes
but those are only valid in strings which means that there would be
other text as well which then makes the removal of this check pointless.