Add missing field check to jinfochk and jauthchk
Although the satirical member

    "no-comment" : "mandatory JSON parsing directive"

(or JSON_PARSING_DIRECTIVE_NAME : JSON_PARSING_DIRECTIVE_VALUE)

was added to the common fields list the field was not handled in the
check function. This did not trigger an error in the calls to jinfochk
and jauthchk because this was not declared an issue as it suggests a
field is not handled rather than it necessarily being an invalid file.

This check was added and I've updated the comment on that point to have:

    * XXX On the other hand it can result in a false positive of a
    * valid test so this has to be carefully considered. However
    * since the jinfochk and jauthchk tools will change
    * dramatically this function might not even exist (or if it
    * does it'll be very different). Either the idea that an
    * unhandled field is not an issue with the file could be
    * reassessed even if an unhandled field is actually a problem
    * with the tools themselves. This same logic applies to the
    * .info.json and .author.json fields check functions.

as it might be a problem after all - it would allow an out of date
function to be incorrect. As noted though these functions might very
well be removed but in any event it'll be done quite differently (unless
some conversion code is written for the struct json * to these structs)
but this is all dependent on the JSON parser being finished.