ioccc_test.sh now runs jparse_test.sh
Exit codes 9 and 10 have been used: 9 for the jparse_test.sh not
existing or not being an executable and 10 for the json_teststr.txt file
not being a readable file.

At the same time I've changed the 'not a file' to 'not a regular file'
(this might need to be done in some of the other tools too) as it is at
this point already established that it's a file of some kind.

I also updated comments for each firewall check to note what is being
checked just to segmentalise (yes that is a word) it a bit better/more.