Add missing files to jparse.1
Specifically json_{parse,util}.{h,c}.

A question remains of whether or not the reference files should be added
to the man page (and the files generated by flex and bison in general).
There is an argument for and against: on the one hand it is part of the
tool; but on the other hand they shouldn't be touched. Even so it might
be better as that's what is actually compiled to make the tool. But this
can be decided upon and if necessary done another time.