Add files to jparse.1
The following files were added:

- jparse.tab.h
- jparse.tab.c
- jparse.c
- sorry.tm.ca.h

For the time being at least this should be all the files related to
jparse directly except perhaps json.h and json.c.

However it might be an idea to add the other json files as well as
they're also still related to jparse even if indirectly so.