Fix some comments in jparse.l
The parsing of grammar should be correct (though there are some that are
missing) but it does not build a parse tree. The previous statement
suggesting that parsing is not correct has been removed.

Also the fact the parse_json() and parse_json_file() functions only warn
on error is brought up as something to reconsider because when a
conversion function fails to parse something it actually is made fatal.
Even so it is useful to have the ability to parse more than one file or
string for now. A possible way to go about it is to use the werr()
function which is like err() but more a warning - it's to be an error
but it's not a fatal error. Currently this is not done but it is
something that might be considered (or else change them all to err()).