Define JSON_BYTE_VALUES to be (BYTE_VALUES)
In json code (e.g. for jenc table) use JSON_BYTE_VALUES to make it a bit
clearer the purpose of the value. If for some reason BYTE_VALUES ever
changes so too will JSON_BYTE_VALUES as it's done as:

    #define JSON_BYTE_VALUES (BYTE_VALUES)

But this way it's clearer that it's meant for JSON. There's a reference
to BYTE_VALUES still in util.c for the hexval table.

I also added notes to the top of jenc table to say that JSON_BYTE_VALUES
is the same as (BYTE_VALUES) and that the table MUST be 256 elements in
size (and this is verified in jencchk()).