Add run_{flex,bison}.sh to README.md
There currently are no man pages for these tools so these have not been
referenced yet. I have started the man page for run_flex.sh but it will
have to be finished another time.

Because of formatting (specifically the '_' in the names) I surrounded
the name of each tool (in each section title after ##) with `s as well.