Many dbg.c and dbg.md updates
The changes are for consistency (but see below) as well as typo fixes
and trying to reduce the length of the text (removing some redundancy).
Some points of interest:

In most cases the texts describing what happens if fmt == NULL or name
== NULL have been moved to the top of the dbg.md to remove duplicate
text. The only difference between each instance was the function name
but I used 'foo' and put below that 'foo' is whichever function is being
used.

Reworded some things to help make said things clearer.

Added missing words in some sentences.

In the example code I changed the VERSION_STRING to be the current date
(or actually 1 June as I made that change yesterday). This is because
the date referred to 2021 and many improvements have been made since
2021. Maybe when everything is in order the date should be updated to
the date the texts are all in order? I don't know.

I got rid of some superfluous commas. More can be done here. However I
have tried to keep the format the same as the original document. If I
encounter any Oxford commas I'll try keeping those in as well even
though it goes against my nature. Important point is I think commas are
overused in much text out there but I am trying to keep them in here to
match the original style.

There's much left to do. I've not yet done a full pass in either file
and this will probably take several complete passes.

As for consistency: some of the things I have updated for this are not
entirely consistent as I have not gone through either file completely
once. I wanted to make this commit to have another clean start.