Remove space at end of line in dbg.md
This was accidentally added in commit
1997f3fc7eab946c13e9684feecfcc44241c5f17.

A vim tip to help prevent this problem:

I'll take this as an opportunity to give this tip for those who use vim.
If you put in your .vimrc file the following for say .md files
(including even a file with the exact base name .md):

    autocmd! BufWritePre *.md %s/\s\+$//e

it will remove (upon saving) lines that end with spaces in .md files.
You can do it for any other file type you might edit in vim.

If (like me) you don't want this always enabled but (like me) want it
in certain directories you can put in your .vimrc something like:

    try
	source ~/.vim/override.vim
    catch
	" ignore
    endtry

    try
	source .override.vim
    catch
	" ignore
    endtry

and then put a .override.vim file in the directory with the changes you
want. Just make sure to put the try at the end of the .vimrc file. The
first one is so you can have the same general .vimrc in more than one
system but have system specific overrides.