errno check fix in txzchk + header update
The code:

     errno = 0;
     file->basename = base_name(path);
     if (!file->basename || !strlen(file->basename)) {
	errp(41, __func__, "%s: unable to strdup basename of filename %s", txzpath, path);
 	not_reached();
     }

was problematic because base_name() uses errp() and the final code in
base_name() does not even consider errno. In actuality base_name() calls
errp() on error so the check would have never been reached anyway but
the check is there anyway for extra sanity.

I also moved (in both txzchk.c and txzchk.h) the

     * "Because sometimes even the IOCCC Judges need some help." :-)

so that it's under my name as it flows a bit better: the reason it was
written by me is because sometimes even the IOCCC Judges need some help.