Add test suite for txzchk - issue #235
    usage: ./txzchk_test.sh [-h] [-v level] [-D dbg_level] [-t txzchk] [-T tar]
			 [-F fnamchk] [-d txzchk_tree]

	-h				print help and exit 2
	-v level			set verbosity level for this script: (def level: 0)
	-D dbg_level			set verbosity level for tests (def: level: 0)
	-t /path/to/txzchk		path to txzchk executable (def: ./txzchk)
	-T /path/to/tar			path to tar that accepts -J option
	-F /path/to/fnamchk		path to fnamchk (def: ./fnamchk)

	-d txzchk_tree			tree where txzchk test files are to be found (def: ./test_txzchk)
					These subdirectories are expected:
					    txzchk_tree/bad
					    txzchk_tree/good

    exit codes:
	0 - all is well
	1 - at least one test failed
	2 - help mode exit
	3 - invalid command line
	>= 4 - internal error

This is very much like json_test.sh except only one tool is checked and
there's no need for a code checker either. In general text files need to
be added. However as discussed there are a couple tarballs in each of
test_txzchk/good and test_txzchk/bad.  No more tarballs need to be added
here (though a few might be); many more text files should be added. This
will help find faults in the tool.

See test_txzchk/README.md for details on the format of these files. More
could be said there; I can elaborate later if necessary.

As for the script itself there are two functions that matter:
run_txz_test which runs txzchk on tarballs and run_txt_test which runs
txzchk on text files. The key difference is that for the text file the
following has to be appended to the command line prior to the --:

     -T -E txt

This cannot be changed! The -T says to not use tar but rather read it as
a text file (you can also read it as it's test mode as it's primarily
for testing new formats and validity of tar output without having to
construct tarballs: that's what I originally implemented it for from
almost the very beginning). However the -E is needed because it modifies
the behaviour of fnamchk: otherwise fnamchk would break and it would
mean that txzchk would fail. Note that files ending in .txz.txt are not
allowed either! The default extension is txz so there's no need to
specify -E txz in the run_txz_test.

An important note on adding test files and .gitignore is that it has the
following glob:

    entry.*.tx[zt]

Which would mean -f is required to `git add'. This is important but
under test_txzchk this is negated so one should not need to add -f to
`git add` when adding to test_txzchk.

One more thing is that just like json_test.sh there is a problem in that
for some reason even if a test fails the script reports that everything
passed. I've not tried to figure out why it is in the other script and
so it's also in this script.

With the exception of this above mentioned problem and adding more test
files this resolves issue #235.