Fixed issue #238
Issue #238:

    Bug: json_test.sh always exits 0 even when EXIT_CODE is 1

that also impacted `txzchk_test.sh` has been fixed.

The problem was that `EXIT_CODE` was being set in a sub-shell
and so the parent shell never saw the non-zero `EXIT_CODE` value.

While there are may work-a-rounds for this issue, we chose
to use a temporary file to hold sub-shell exit codes.
The parent shell creates the temporary exit code file,
the sub-shell writes a numeric value into that same
temporary exit code file, so that in the end, the parent
shell can inspect the temporary exit code file and
exit accordingly.

Some level of care is taken in using temporary exit code files.
A mktemp template filename is used that includes a common prefix.
The value in the temporary exit code file must be an integer
in the range [0,255].  The final EXIT_CODE, if 0, is replaced
by the validated exit code read from the temporary exit code file.
In the unlikely event where the temporary exit code file contains 0,
the final EXIT_CODE is not changed.

Continuing the **Yak Shave**, use of the `mktemp` command was
improved in other shell scripts that use it, namely `run_bison.sh`
and `run_flex.sh`.  They too use a mktemp template filename.

Temporary files created by `json_test.sh` and `txzchk_test.sh`
are of the form: `.exit_code.__script_basename__.XXXXXXXXXX*`
and the temporary files created by `run_bison.sh` are of the
form `.sorry.bison.XXXXXXXXXX` and temporary files created by
``run_flex.sh` are of the form `.sorry.flex.XXXXXXXXXX`.

All temporary exit code files are cleaned up by the `make prep_clobber`
rule.  They are also ignored in the `.gitignore` file.

Fixed a problem where the `build.log` file was being
removed during the `make prep` and `make release` rules.

A non-zero exit of `prep.sh` during `make prep` and `make release`
will be reported by the Makefile.  In particular, if `prep.sh`
exited non-zero, then a Warning is issued by the make file.  In the
case of `make prep`, the make command will not exit.  In the case
of `make release`, the make command exit non-zero (abort) after
printing information such as the `prep.sh` exit code and a message
about above details were saved in the build.log file.