Add bad test files in test_txzchk/bad + typo fix
Add the following test files to test_txzchk/bad (txzchk must report at
least one issue for each file - which it does and txzchk_test.sh passes
all tests):

    entry.test-0.1652598666.txt
    entry.test-0.9876543210.txt
    entry.test-1.9876543210.txt
    entry.test-10.9876543210.txt
    entry.test-2.1652598666.txt
    entry.test-2.9876543210.txt
    entry.test-3.1652598666.txt
    entry.test-3.9876543210.txt
    entry.test-4.1652598666.txt
    entry.test-4.9876543210.txt
    entry.test-5.1652598666.txt
    entry.test-5.9876543210.txt
    entry.test-6.1652598666.txt
    entry.test-6.9876543210.txt
    entry.test-7.9876543210.txt
    entry.test-9.9876543210.txt

These files are based on the tarballs that were used during development
of txzchk to make sure all issues that txzchk had to detect were in fact
detected. The only difference is that they're now text files and the
timestamp was updated as the MIN_TIMESTAMP has been updated since the
time of the creation of txzchk.

As for the file:

    entry.test-10.9876543210.txt

A point of interest might be that fnamchk will fail so txzchk will
report that the directory name is indeed wrong. You would see an error
like so:

    ERROR[6]: main: second '-' separated token length: 7 != 6: ./test_txzchk/bad/entry.test-10.9876543210.txt

This is because of:

    #define MAX_ENTRY_CHARS (1)             /* characters that represent the maximum entry number */

which is used in fnamchk as:

        if (len != LITLEN("test-")+MAX_ENTRY_CHARS) {
	    /* ... */
	}

and

        if (len != UUID_LEN+1+MAX_ENTRY_CHARS) {
	    /* ... */
	}

Whether fnamchk should print the directory name at this point and then
return an error is I suppose up to discussion: it would allow for
testing more but it might be unnecessary as well given that a badly
named tarball would presumably be disqualified. Even so it might help
uncover some types of problems with txzchk.

TODO add even more tarball listings. I am pretty sure I have some that I
created as well but more can be contrived as well (one or more of the
above might actually be some I created myself but all the invalid
tarballs that were used for development are also included).

As for the typo fix: in txzchk_test.sh it was in a variable name:
EXIT_CODE_FIlE -> EXIT_CODE_FILE.