Typo fixes and shortening usage in txzchk_test.sh
Comment at top referred to json_test.sh related (as I based it on that
script as it is similar in nature in that it expects good files to pass
and bad files to fail).

Also there's no need to have '/path/to/foo' when the description in the
usage (to the right of /path/to/foo') says it's the path to foo. There
probably are other tools that have this same issue but those can be
fixed another time.