Initial dbg.3 man page
This is not likely complete but it's good start for initial analysis and
discussion.

More examples should be added (from each section in the dbg.md file).

Formatting has to be checked in its entirety. An example is in C code.
Instead of having - one has to have \- (even in C code). I believe I got
all cases but I might have missed some. Similarly it appears that \n in
C should be in the man page \\n.

Consistency should be checked more thoroughly as well. A good example is
making sure that the va_list functions are in the same order as the
regular functions when indicating equivalency (other than taking a
va_list). Another example of consistency that needs to be checked is the
order of the functions in the man page versus dbg.h.

The dbg.md does not yet refer to the man page and won't until it is
completed and made as concise and precise as possible.