Set verbosity_level in dbg.md main example
This was also done in dbg.3 and that's what prompted me to do it here as
well.

Something I neglected to mention in commit
881b1970a8c291180693d9e1d6b86552eaf476ec is that there's no explanation
about compiling dbg.c and linking dbg.o into the binaries that wish to
use the dbg facility. Where and how this should be displayed is not
something I am sure of yet.