First complete pass of dbg.3 man page
The man page is now ready for a thorough review and if everything checks
out okay the dbg.md file could be updated to refer to it. It's possible
another pass over the man page would be a good idea though. This
specifically goes for what I noted in commit
881b1970a8c291180693d9e1d6b86552eaf476ec. I believe I can tick these but
they should still be checked nonetheless:

    Formatting has to be checked in its entirety. An example is in C code.
    Instead of having - one has to have \- (even in C code). I believe I got
    all cases but I might have missed some. Similarly it appears that \n in
    C should be in the man page \\n.

Note that in this case the dbg_example.c file itself obviously cannot
have the '\'s in front of '-'s and '\'s.

What I cannot say can be ticked yet is:

    Consistency should be checked more thoroughly as well. A good example is
    making sure that the va_list functions are in the same order as the
    regular functions when indicating equivalency (other than taking a
    va_list). Another example of consistency that needs to be checked is the
    order of the functions in the man page versus dbg.h.

For the first part I have not done it at all though for writing it in
the first place I made each function on its own line and then yanked and
pasted and then removed the v from the function name so it should be
okay but I still think it should be checked again.

For the second part I believe I originally had it as well but I could be
wrong as I did go about doing it in different orders at first.

What can now likely be discussed is this:

    The dbg.md does not yet refer to the man page and won't until it is
    completed and made as concise and precise as possible.

Though it might be better to make sure the man page is finished first.
Perhaps not though as reducing the text in dbg.md and referring to the
man page might actually help.

The example in the man page now is a complete program (see
dbg_example.c) and it shows the contents (via cat) and then how to
compile it along with dbg.c. After this it runs './dbg_example' and
shows the output. Finally it shows the exit status of the dbg_example
(which will actually be 2 as I made the final line an errp() call).

For reference this is how to compile the dbg_example:

        $ cc -c dbg.c
        $ cc -o dbg_example dbg_example.c dbg.o

Alternatively you can do:

	$ cc -o dbg_example dbg_example.c dbg.c

A Makefile could have e.g.:

    dbg_example: dbg.o dbg_example.c
	    ${CC} ${CFLAGS} dbg.o dbg_example.c -o $@

or

    dbg_example: dbg.c dbg.h dbg_example.c
	    ${CC} ${CFLAGS} dbg.c dbg_example.c -o $@

for two examples (this list is definitely not exhaustive).

Updated .gitignore to have dbg_example as well as the full range (at
least of lower case - not sure what happens after all the minuscules are
used though I guess it would be [A-Z]) of swap files rather than just
[nop].