Add make checknr (check [nt]roff) and fix *.{1,3}
The checknr tool checks nroff/troff files and this uncovered some
unmatched '\f?'s and also one unmatched .RE (as in there was no opening
.RS).

Since checknr is not on all systems this rule is not run by any of the
check rules as this would mean any system that does not have it
installed would fail.

The rule looks like:

    # inspect and verify man pages
    #
    checknr: ${MANPAGES}
	    @HAVE_CHECKNR=`command -v ${CHECKNR}`; if [[ -z "$$HAVE_CHECKNR" ]]; then \
		echo 'The checknr command could not found.' 1>&2; \
		echo 'The checknr command is required to run this rule.'; 1>&2; \
		echo ''; 1>&2; \
		exit 1; \
	    else \
		echo "${CHECKNR} ${MANPAGES}"; \
		${CHECKNR} ${MANPAGES}; \
	    fi

On the systems I have it was not installed in CentOS or Fedora but it
was installed on macOS. I looked at the package manager on at least one
of those systems and it was not found. Nevertheless it's useful to have
this rule to simplify checking of man pages.