Add request to mention me in issues with txzchk
In txzchk.1 I have added a request to please mention me when reporting
issues. It's unlikely that I won't see it otherwise but hopefully this
will make it even more likely that I will see it. Perhaps not strictly
necessary but I'd rather be certain I see the issue and this should
help.