jparse now prints 'valid json' if valid json
This change involved two steps:

First the parse_json() function at the beginning sets *is_valid = true
if is_valid != NULL. This allows the caller to always have a valid test
result without having to set it to true before each call to the parser.

Second jparse checks the value of the boolean after the parse function
returns and prints a message if it is valid. For files it specifies the
file.

A word about why we don't say if it's invalid. It's because the parse
function itself prints this. It might be good to have the parse function
take a filename and it could specify the file or else jparse could
report invalid as well. For the latter option however it would be
redundant messages as parse_json() actually already reports if it's
invalid.