Add XXX note add_(info|author|common)_json_field
The issue is these functions are named misleadingly. The names would
suggest they add to the (info|author|common)_json_fields tables but they
do not. They actually use the add_found_ versions which add to the
found_ tables. The add_found_ versions should do the work of these
functions instead or else these should be renamed to something else
that's not misleading (and more so more clear).