Typo fixes
In run_bison.sh and run_flex.sh in add_sorry it said 'expects 1 args'
instead of 'arg'.

Why not change it to 'one'? It might be up to debate but the reason I
did not is because $# shows digits and not words so I left it as digits
too to keep it matching.

In util.h: I moved a comment above

    -/* for string to int functions */
    -#define LLONG_MAX_BASE10_DIGITS (19)
    +#define LLONG_MAX_BASE10_DIGITS (19) /* for string to int functions */

because there are lines right below that are not related to string to
int but might be confused as being related in some way or another.