Add initial hostchk.sh
Currently it does not parse options and there are only two checks:

Checks that tar with the required options is available.

Test compiles a C program with each of the system header files included
in all the C and header files in the repo. See below for how this is
done.

Rather than have to update the script each time a header file is removed
or added it uses a pipeline of grep and sort in a while read loop. The
pipeline itself is:

    grep -h -o '#include.*<.*>' ./*.c ./*.h|sort -u

and the top and bottom of the loop is:

    while read -r h; do

    ...

    done < <(grep -h -o '#include.*<.*>' ./*.c ./*.h|sort -u)

But instead of creating a C file for each header file used we use some
magic with the compiler and stdin. Thus the part that compiles each test
file is:

    printf "%s\n%s\n" "$h" "int main(void){ return 0; }" | cc -x c - -o "$PROG_FILE"

Thus there's no need to write to the actual disk as it's all stdin and
stdout.

To run this script use make hostchk or make bug-report.

At the risk of stating the obvious this is not complete.