Error message clean up & fixes in txzchk_test.sh
As brought up at
https://github.com/ioccc-src/mkiocccentry/issues/171#issuecomment-1153018887
the error messages could be a cause for concern in that although they're
expected it might make people think not to report other errors as the
script passes (as it should as these are in test_txzchk/bad).

A special source of errors is fnamchk and this was indeed one of the
ones referred to in the comment. This required some changes to fnamchk
as well. A new -e option has been added to txzchk and fnamchk which
suppresses error messages. The errors still occur but they're not
displayed. Now with txzchk_test.sh if debug level is 0 we use that
option and txzchk in turn will pass -e to fnamchk in that case.

A note is that it only uses -e the first time it calls fnamchk. This is
because the second call in txzchk (where we retrieve the correct
directory) name is only called if fnamchk did not fail so there's no
need to pass -e in this case.

There were some other thoughts that occurred like requiring specific
error messages for specific files but this becomes complicated when one
considers that the fnamchk also checks the MIN_TIMESTAMP which changes
which means that the filenames under test_txzchk/{bad,good} have to be
updated ... which means the script itself would have to be updated each
time.

There also was a thought that the script could show (with some debug
level) which file is about to be processed. This is a good idea but it's
not yet in. That can be done another time as I just wanted to fix this
issue.

The man pages have been updated as well. For fnamchk the -e option was
added; for txzchk the -e option was added and the FILES section was
updated as well: txzchk_test.sh, test_txzchk/bad and test_txzchk/good.