Released dbg 2.2 2022-06-08
Changed dbg facility version to 2.2 2022-06-08.

Added functions to determine if certain debug messages are allowed:

    extern bool msg_allowed(void);
    extern bool dbg_allowed(int level);
    extern bool warn_allowed(void);
    extern bool err_allowed(void);
    extern bool usage_allowed(void);

Now all dbg() functions use the above functions to determine if
messages are allowed to be written.

Probably the most useful is to determine if a given debug level
message is allowed.  For example, say there a complex
debugging action should only be done if the debug level
is the right level, instead of making an explicit check
against the verbosity_level global, a simple call to
to dbg_allowed is needed:

for (i=0; i < bigvalue; ++i) {

    ...

    if (dbg_allowed(DBG_VVHIGH)) {
	dbg(DBG_VVHIGH, "start of special debug section")l
	dbg(DBG_VVHIGH, "i: %ld a: %s b: %d c: %f",
		        i, intensive_a_func(i), intensive_b_func(i), intensive_c_func(i));
	(void) fprintf(dbg_stream, "special message: %s\n", special(i));
	more_complex_debugging(i);
	(void) fprintf(dbg_stream, "very special message: %s\n", very_special(i));
	dbg(DBG_VVHIGH, "end of special debug section")l
    }

    ...
}

By pre-checking if the debug level DBG_VVHIGH is active, calls
to complex debugging functions and even special writes outside
of the dbg facility can be avoided.