Added JSON dbg, warn and error control functions
Added the following JSON dbg, warn and error control functions:

    extern bool json_dbg_allowed(int json_dbg_lvl);
    extern bool json_warn_allowed(int code);
    extern bool json_err_allowed(void);

The json_warn_allowed() was written to be similar to
the warn_allowed() function, except that the warning
code is passed to json_warn_allowed()

The json_err_allowed() was written to be similar to
the err_allowed() function.

Moved show_full_json_warnings global up near the
top of the json_util.c file.

Renamed level arg to json_dbg_lvl.