Simplified JSON messaging interface
The program name and debugging streams have been removed
as arguments to many JSON functions.

The program name remains used in main() and the related usage()
message.

It became too awkward to try can carry around as arguments
to various JSON functions, the program name and the debugging stream.

The program name, if it is needed, is best established in main()
while the command line arguments are being processed.  Low level
debugging, on the other hand, can be deep inside functions.  To
pass around the program name as a arg becomes too complex.

If there is a need for the program name, it is probably best to
access it as a global.  The program name is one of the few useful
global values to have, even in a multi-threaded application as all
threads are part of the same application.  However requiring the
use of a program global requires the program calling the parser
to conform to that notion, complicating the interface.

Carrying around a debugging stream as an argument is also problematic
for similar reasons.  Moreover the debugging stream is almost often
stderr.  If the application wants such messages appear elsewhere,
stderr can be redirected to that elsewhere either on the command
line, or via calling freopen() on stderr.

The dbg facility does have functions for specifying the stream,
however the parallel json dbg facility did not.  It seemed better
to remove the notion of dbg_stream complete than to either build
a parallel fjson_foo() function interface like the dbg facility
has and carry around the dbg_stream arg to make use use it.

The JSON parse tree node functions are able to carry an output
stream within themselves.  This is to allow writing to stdout if
needed.  However the top level json_dbg_tree_print() function,
because it is a debugging function that may be used deep internals,
uses stderr and calls internally the JSON parse tree node functions
with an explicit stderr steam value.

To recap: The program name is under the control of main() and
the command line parsing related functions that are closely
tied to main().

To recap: debugging information should go to stderr.  If such
debugging is needed to be put elsewhere, then stderr can be
redirected accordingly.