Improve JSON debug levels
By default, the ugly_debug is set to 0; verbose bison debug off.

When -J 7 or higher is used, ugly_debug is set to 1 and verbose
bison debug is set to on.

Fixed a few issues with missing newlines in fprstr() and fprint()
output.

Disabled flex debug by default.

Flush stderr, in case it is buffered, in ugly_error().