Many typo fixes in scripts and source files
There were many references to 'not a file' after first determining that
it actually exists and is a file (should have read 'not a regular
file').

In json_test.sh a variable was called EXIT_CODE_FIlE instead of
EXIT_CODE_FILE (if it's not obvious at a quick glance the former name
had a lower case ell).

The C files:

   jauthchk.c
   jinfochk.c
   mkiocccentry.c
   txzchk.c

All referred to 'not a file' after first determining that said path is
in fact a file (again here should have read 'not a regular file' as they
now do).

In vermod.sh change the name (in comment) to vermod.sh (was vermod).

There might have been some other fixes as well but the above are the
significant ones.