Change lambda transliterations in utf8_posix_map.c
According to OED lambda is:

    the eleventh letter of the Greek alphabet (Λ, λ), transliterated as ‘l’.

so it should be capital (or uncial; majuscule) and lower case (or
minuscule) ELL and not y and n. Thus:

    -    { "\xc6\x9b", "y" , -1, -1},	/*  U+019B - ƛ - LATIN SMALL LETTER LAMBDA WITH STROKE */
    +    { "\xc6\x9b", "l" , -1, -1},	/*  U+019B - ƛ - LATIN SMALL LETTER LAMBDA WITH STROKE */

    -    { "\xce\x9b", "n" , -1, -1},	/*  U+039B - Λ - GREEK CAPITAL LETTER LAMDA */
    +    { "\xce\x9b", "L" , -1, -1},	/*  U+039B - Λ - GREEK CAPITAL LETTER LAMBDA */

    -    { "\xce\xbb", "y" , -1, -1},	/*  U+03BB - λ - GREEK SMALL LETTER LAMDA */
    +    { "\xce\xbb", "l" , -1, -1},	/*  U+03BB - λ - GREEK SMALL LETTER LAMBDA */

Now the question of the Latin version. As best I could find this is not
even actually a thing other than in Unicode. I believe the Greek lambda
created the Latin L though so it seems appropriate that the change be
made for the Latin version too.

Also I fixed the spelling 'lamda' to 'lambda'. Apparently Unicode uses
'lamda' for Greek and lambda for Latin which is both ironic and
ridiculous as lambda is the actual letter in the Greek alphabet and it's
spelt lambda not 'lamda' as well.

If there's a 'Latin capital letter lambda with stroke' it probably
should be changed to 'L' as well. However I don't see it in the table so
I'm guessing there isn't anything left to do here.